# Triage-app/backend/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.PhysioCallRequest:
    properties:
      call_type:
        description: Immediate or Scheduled
        type: string
      initiated_by:
        type: string
      scheduled_time:
        description: Optional for immediate calls
        type: string
    required:
    - call_type
    - initiated_by
    type: object
...
  services.PhysioCall:
    properties:
      assessment_id:
        type: string
      call_id:
        type: string
      call_status:
        description: Scheduled, Completed, Cancelled
        type: string
      call_type:
        description: Immediate or Scheduled
        type: string
      created_at:
        type: string
      initiated_by:
        description: User or System
        type: string
      scheduled_time:
        description: NULL for immediate calls
        type: string
    type: object
...
  services.ROMAnalysis:
    properties:
      analysis_result:
        type: string
      assessment_id:
        type: string
      created_at:
        type: string
      pose_model_data:
        description: JSONB type for pose model data
    type: object
  services.SelfCarePlan:
    properties:
      assessment_id:
        type: string
      created_at:
        type: string
      critical_flag:
        type: boolean
      plan_name:
        type: string
      suggested_exercises:
        description: JSONB for exercises
    type: object
host: localhost:8080
info:
  contact: {}
  description: API documentation for AI-Bot-DeeCogs
  title: AI-Bot-DeeCogs API
  version: "1.0"
paths:
...
  /assessments/:id/physio-calls:
    get:
      description: Retrieves all physio calls for a specific assessment
      parameters:
      - description: Assessment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/services.PhysioCall'
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get physio calls
      tags:
      - Physio Calls
...
  /assessments/{id}/physio-calls:
    post:
      consumes:
      - application/json
      description: Schedules a physio call (immediate or scheduled) for a specific
        assessment
      parameters:
      - description: Assessment ID
        in: path
        name: id
        required: true
        type: string
      - description: Call Details
        in: body
        name: call_data
        required: true
        schema:
          $ref: '#/definitions/handlers.PhysioCallRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/services.PhysioCall'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Schedule a physio call
      tags:
      - Physio Calls
  /assessments/{id}/rom:
    get:
      description: Retrieves pose model data and analysis results for an assessment
      parameters:
      - description: Assessment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/services.ROMAnalysis'
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get ROM analysis data
      tags:
      - ROM
    post:
      consumes:
      - application/json
      description: Submits pose model data and analysis results for an assessment
      parameters:
      - description: Assessment ID
        in: path
        name: id
        required: true
        type: string
      - description: ROM Data
        in: body
        name: rom_data
        required: true
        schema:
          $ref: '#/definitions/services.ROMAnalysis'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Submit ROM analysis data
      tags:
      - ROM
  /assessments/{id}/self-care-plans:
    get:
      description: Retrieves personalized self-care plans for an assessment
      parameters:
      - description: Assessment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.SelfCarePlan'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get self-care plans
      tags:
      - Self-Care Plans
...
  /users:
    post:
      consumes:
      - application/json
      description: Creates a new user with the provided details
      parameters:
      - description: User Data
        in: body
        name: user
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new user
      tags:
      - Users
  /users/{id}:
    get:
      description: Retrieves details of a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user details
      tags:
      - Users
swagger: "2.0"